feat: comprehensive SEO optimization and auth forms refactoring
- Migrated auth forms to centralized auth-forms folder for better organization
- Added comprehensive SEO metadata and structured data
- Implemented dynamic sitemap generation with proper URLs
- Enhanced robots.txt with proper crawling rules
- Added page-specific metadata for all auth and dashboard pages
- Protected private pages from search engine indexing
- Added OpenGraph and Twitter card optimizations
- Implemented structured data for Website, Organization, and WebApplication
- Enhanced mobile viewport configuration with theme color support
- Optimized font loading with display swap for better performance
- Created reusable SEO utilities for consistent metadata across pages
