+++ .gitignore
@@ -41,3 +41,6 @@ yarn-error.log*
 next-env.d.ts
 
 .env
+.env*.local
+
+todo.md
