--- public/robots.txt
@@ -1,11 +1,0 @@
-User-agent: *
-Allow: /
-
-# Sitemap
-Sitemap: https://starter.nabinkhair.com.np/sitemap.xml
-
-# Disallow admin and API routes
-Disallow: /api/
-Disallow: /admin/
-Disallow: /_next/
-Disallow: /static/
