+++ src/app/(auth)/components/auth-forms/index.ts
@@ -1,0 +1,5 @@
+export { CreateAccountForm } from './create-account';
+export { EmailVerificationForm } from './email-verification';
+export { ForgotPasswordForm } from './forgot-password';
+export { LoginForm } from './login';
+export { ResetPasswordForm } from './reset-password';
