+++ src/app/(auth)/create-account/page.tsx
@@ -1,238 +1,24 @@
-'use client';
+import { CreateAccountForm } from '@/app/(auth)/components/auth-forms';
+import { generatePageMetadata } from '@/seo/seo';
 
-import { Button } from '@/components/ui/button';
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from '@/components/ui/card';
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from '@/components/ui/form';
-import { Input } from '@/components/ui/input';
-import { Separator } from '@/components/ui/separator';
-import { useCreateAccount } from '@/hooks/auth/useAuthMutations';
-import { createAccountSchema } from '@/zod/authValidation';
-import { zodResolver } from '@hookform/resolvers/zod';
-import { Eye, EyeOff, User } from 'lucide-react';
-import { useRouter } from 'next/navigation';
-import { useState } from 'react';
-import { useForm } from 'react-hook-form';
-import { z } from 'zod';
-import { OAuthButtons } from '../oauth-buttons';
-type CreateAccountFormData = z.infer<typeof createAccountSchema>;
+export const metadata = generatePageMetadata({
+  title: 'Create Account',
+  description:
+    'Create a new SaaS Starter account to get started with our modern SaaS platform. Sign up in seconds and access all features.',
+  keywords: [
+    'create account',
+    'sign up',
+    'register',
+    'new user',
+    'registration',
+  ],
+  canonical: '/create-account',
+});
 
 export default function CreateAccountPage() {
-  const [showPassword, setShowPassword] = useState(false);
-  const [showConfirmPassword, setShowConfirmPassword] = useState(false);
-  const router = useRouter();
-  const createAccountMutation = useCreateAccount();
-
-  const form = useForm<CreateAccountFormData>({
-    resolver: zodResolver(createAccountSchema),
-    defaultValues: {
-      firstName: '',
-      lastName: '',
-      email: '',
-      password: '',
-      confirmPassword: '',
-    },
-    mode: 'onChange',
-  });
-
-  const onSubmit = async (data: CreateAccountFormData) => {
-    if (data.password !== data.confirmPassword) {
-      form.setError('confirmPassword', {
-        message: 'Passwords do not match',
-      });
-      return;
-    }
-
-    try {
-      await createAccountMutation.mutateAsync({
-        ...data,
-        confirmPassword: data.confirmPassword || '',
-      });
-      // Store email in localStorage for email verification page
-      localStorage.setItem('verificationEmail', data.email);
-      router.push('/email-verification');
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
   return (
     <div className="flex items-center justify-center">
-      <Card className="w-full max-w-md">
-        <CardHeader>
-          <div className="mb-4 flex h-12 w-12 items-center justify-center rounded-xl bg-muted">
-            <User className="h-6 w-6" />
-          </div>
-          <CardTitle className="text-2xl font-bold">Create Account</CardTitle>
-          <CardDescription>
-            Enter your information to create a new account
-          </CardDescription>
-        </CardHeader>
-        <CardContent>
-          <Form {...form}>
-            <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4">
-              <div className="grid grid-cols-2 gap-4">
-                <FormField
-                  control={form.control}
-                  name="firstName"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>First Name</FormLabel>
-                      <FormControl>
-                        <Input placeholder="Enter your first name" {...field} />
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-                <FormField
-                  control={form.control}
-                  name="lastName"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Last Name</FormLabel>
-                      <FormControl>
-                        <Input placeholder="Enter your last name" {...field} />
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-              </div>
-
-              <FormField
-                control={form.control}
-                name="email"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Email</FormLabel>
-                    <FormControl>
-                      <Input
-                        type="email"
-                        placeholder="Enter your email"
-                        {...field}
-                      />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-
-              <FormField
-                control={form.control}
-                name="password"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Password</FormLabel>
-                    <FormControl>
-                      <div className="relative">
-                        <Input
-                          type={showPassword ? 'text' : 'password'}
-                          placeholder="Enter your password"
-                          {...field}
-                        />
-                        <Button
-                          type="button"
-                          variant="ghost"
-                          size="sm"
-                          className="absolute right-2 top-1/2 -translate-y-1/2 h-6 w-6 p-0 hover:bg-transparent"
-                          onClick={() => setShowPassword(!showPassword)}
-                        >
-                          {showPassword ? (
-                            <EyeOff className="h-4 w-4" />
-                          ) : (
-                            <Eye className="h-4 w-4" />
-                          )}
-                        </Button>
-                      </div>
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-
-              <FormField
-                control={form.control}
-                name="confirmPassword"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Confirm Password</FormLabel>
-                    <FormControl>
-                      <div className="relative">
-                        <Input
-                          type={showConfirmPassword ? 'text' : 'password'}
-                          placeholder="Confirm your password"
-                          {...field}
-                        />
-                        <Button
-                          type="button"
-                          variant="ghost"
-                          size="sm"
-                          className="absolute right-2 top-1/2 -translate-y-1/2 h-6 w-6 p-0 hover:bg-transparent"
-                          onClick={() =>
-                            setShowConfirmPassword(!showConfirmPassword)
-                          }
-                        >
-                          {showConfirmPassword ? (
-                            <EyeOff className="h-4 w-4" />
-                          ) : (
-                            <Eye className="h-4 w-4" />
-                          )}
-                        </Button>
-                      </div>
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-
-              <Button
-                type="submit"
-                className="w-full"
-                disabled={createAccountMutation.isPending}
-                loading={createAccountMutation.isPending}
-                loadingText="Creating"
-              >
-                Create Account
-              </Button>
-            </form>
-          </Form>
-          <div className="my-6 flex items-center gap-4">
-            <Separator className="flex-1" />
-            <span className="text-sm text-muted-foreground">
-              Or sign in with
-            </span>
-            <Separator className="flex-1" />
-          </div>
-
-          <OAuthButtons />
-        </CardContent>
-        <CardFooter className="flex justify-center">
-          <div className="text-center text-sm">
-            Already have an account?{' '}
-            <Button
-              variant="link"
-              className="p-0 font-normal"
-              onClick={() => router.push('/log-in')}
-            >
-              Sign in
-            </Button>
-          </div>
-        </CardFooter>
-      </Card>
+      <CreateAccountForm />
     </div>
   );
 }
