+++ src/app/sitemap.ts
@@ -1,0 +1,6 @@
+import { generateSitemap } from '@/seo';
+import { MetadataRoute } from 'next';
+
+export default function sitemap(): MetadataRoute.Sitemap {
+  return generateSitemap() as MetadataRoute.Sitemap;
+}
