+++ src/components/OAuthCallback.tsx
@@ -68,10 +68,12 @@ export const OAuthCallback = () => {
   }, [searchParams, router, login]);
 
   return (
-    <div className="flex items-center justify-center min-h-screen">
+    <div className="flex items-center justify-center min-h-[100svh]">
       <div className="text-center">
         <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-gray-900 mx-auto"></div>
-        <p className="mt-4 text-gray-600">Completing authentication...</p>
+        <p className="mt-4 text-muted-foreground">
+          Completing authentication...
+        </p>
       </div>
     </div>
   );
