+++ src/components/ProtectedRoute.tsx
@@ -33,7 +33,7 @@ export default function ProtectedRoute({
   if (isLoading) {
     return (
       fallback || (
-        <div className="flex min-h-screen items-center justify-center bg-background">
+        <div className="flex min-h-[100svh] items-center justify-center bg-background">
           <Loader className="animate-spin" size={24} />
         </div>
       )
