+++ src/components/screen-loader.tsx
@@ -1,9 +1,8 @@
-import React from 'react';
 import { Loader2 } from 'lucide-react';
 
 const ScreenLoader = () => {
   return (
-    <div className="flex items-center justify-center h-screen">
+    <div className="flex items-center justify-center h-[100svh]">
       <Loader2 className="animate-spin" />
     </div>
   );
