+++ src/seo/seo.ts
@@ -1,0 +1,128 @@
+import { siteConfig } from '@/seo/metadata';
+import { Metadata } from 'next';
+
+interface SEOProps {
+  title?: string;
+  description?: string;
+  keywords?: string[];
+  canonical?: string;
+  ogImage?: string;
+  noIndex?: boolean;
+}
+
+export function generatePageMetadata({
+  title,
+  description,
+  keywords = [],
+  canonical,
+  ogImage,
+  noIndex = false,
+}: SEOProps = {}): Metadata {
+  const metaTitle = title ? `${title} | ${siteConfig.name}` : siteConfig.name;
+  const metaDescription = description || siteConfig.description;
+  const metaKeywords = [...(keywords || []), 'SaaS', 'Next.js', 'TypeScript'];
+  const metaImage = ogImage || siteConfig.ogImage;
+
+  return {
+    title: metaTitle,
+    description: metaDescription,
+    keywords: metaKeywords,
+    robots: noIndex
+      ? { index: false, follow: false }
+      : { index: true, follow: true },
+    alternates: canonical ? { canonical } : undefined,
+    openGraph: {
+      title: metaTitle,
+      description: metaDescription,
+      url: canonical || siteConfig.url,
+      images: [
+        {
+          url: metaImage,
+          width: 1200,
+          height: 630,
+          alt: metaTitle,
+        },
+      ],
+      type: 'website',
+      siteName: siteConfig.name,
+    },
+    twitter: {
+      card: 'summary_large_image',
+      title: metaTitle,
+      description: metaDescription,
+      images: [metaImage],
+      creator: '@nabinkhair42',
+    },
+  };
+}
+
+// Structured data generators
+export function generateWebsiteStructuredData() {
+  return {
+    '@context': 'https://schema.org',
+    '@type': 'WebSite',
+    name: siteConfig.name,
+    url: siteConfig.url,
+    description: siteConfig.description,
+    potentialAction: {
+      '@type': 'SearchAction',
+      target: `${siteConfig.url}/search?q={search_term_string}`,
+      'query-input': 'required name=search_term_string',
+    },
+  };
+}
+
+export function generateOrganizationStructuredData() {
+  return {
+    '@context': 'https://schema.org',
+    '@type': 'Organization',
+    name: 'SaaS Starter',
+    url: siteConfig.url,
+    logo: `${siteConfig.url}/icon.png`,
+    description: siteConfig.description,
+    sameAs: [siteConfig.links.github, siteConfig.links.twitter],
+    founder: {
+      '@type': 'Person',
+      name: 'Nabin Khair',
+      url: 'https://nabinkhair.com.np',
+    },
+  };
+}
+
+export function generateWebApplicationStructuredData() {
+  return {
+    '@context': 'https://schema.org',
+    '@type': 'WebApplication',
+    name: siteConfig.name,
+    url: siteConfig.url,
+    description: siteConfig.description,
+    applicationCategory: 'BusinessApplication',
+    operatingSystem: 'Web Browser',
+    offers: {
+      '@type': 'Offer',
+      price: '0',
+      priceCurrency: 'USD',
+      category: 'Freemium',
+    },
+    creator: {
+      '@type': 'Organization',
+      name: 'SaaS Starter',
+      url: siteConfig.url,
+    },
+  };
+}
+
+export function generateBreadcrumbStructuredData(
+  items: Array<{ name: string; url: string }>
+) {
+  return {
+    '@context': 'https://schema.org',
+    '@type': 'BreadcrumbList',
+    itemListElement: items.map((item, index) => ({
+      '@type': 'ListItem',
+      position: index + 1,
+      name: item.name,
+      item: item.url,
+    })),
+  };
+}
