+++ todo.md
@@ -1,0 +1,7 @@
+1. Optimize the code and remove redundant code
+2. Improve the performance
+3. Improve the /dashboard route details
+4. Think of Landing Page
+5. Add few more features like: Last Used badge in the login screen
+6. Add Magic Based Login
+7. Improve the Mail Templates UX
